Store capabilities in unsigned var

+++ xine_post_autocrop.c
@@ -4,7 +4,7 @@
  * See the main source file 'xineliboutput.c' for copyright information and
  * how to reach the author.
  *
- * $Id: xine_post_autocrop.c,v 1.40 2010-05-22 12:15:39 phintuka Exp $
+ * $Id: xine_post_autocrop.c,v 1.41 2010-05-22 12:20:03 phintuka Exp $
  *
  */
 
@@ -1684,7 +1684,7 @@ static post_plugin_t *autocrop_open_plugin(post_class_t *class_gen,
       this->soft_start_step = DEFAULT_SOFT_START_STEP;
       this->stabilize   = 1;
 
-      int caps = port->original_port->get_capabilities(port->original_port);
+      uint64_t caps = port->original_port->get_capabilities(port->original_port);
       this->has_driver_crop = caps & VO_CAP_CROP;
       this->has_unscaled_overlay = caps & VO_CAP_UNSCALED_OVERLAY;
